used initBinConfig and compute_bins instead of duplicating logic
