+++ server/shared/types/dataset.ts
@@ -316,6 +316,7 @@ export type GeneExpressionQueryNative = {
 	samples?: number[]
 	nochr?: boolean
 	get?: (param: any) => void
+	//This dictionary is used to store/cache the default bins calculated for a geneExpression term when initialized in the fillTermWrapper
 	gene2bins?: { [index: string]: any }
 }
 export type GeneExpressionQuery = GeneExpressionQueryGdc | GeneExpressionQueryNative
