+++ server/src/termdb.js
@@ -7,7 +7,7 @@ import { validate as snpValidate } from './termdb.snp.js'
 import { isUsableTerm } from '#shared/termdb.usecase.js'
 import { trigger_getSampleScatter } from './termdb.scatter.js'
 import { trigger_getLowessCurve } from './termdb.scatter.js'
-import { getData, getSamplesPerFilter, getDefaultGeneExpBins } from './termdb.matrix.js'
+import { getData, getSamplesPerFilter } from './termdb.matrix.js'
 import { get_mds3variantData } from './mds3.variant.js'
 import { get_lines_bigfile, mayCopyFromCookie } from './utils.js'
 import { authApi } from './auth.js'
@@ -15,6 +15,9 @@ import { getResult as geneSearch } from './gene.js'
 import { searchSNP } from '../routes/snp.ts'
 import { get_samples_ancestry, get_samples } from './termdb.sql.js'
 import { TermTypeGroups } from '#shared/common.js'
+import initBinConfig from '#shared/termdb.initbinconfig'
+import * as bins from '#shared/termdb.bins.js'
+
 /*
 ********************** EXPORTED
 handle_request_closure
@@ -467,6 +470,9 @@ async function LDoverlay(q, ds, res) {
 
 async function trigger_getDefaultGeneExpBins(q, ds, res) {
 	const tw = q.tw
+	if (ds.queries.geneExpression.gene2bins[tw.term.gene])
+		return { default: ds.queries.geneExpression.gene2bins[tw.term.gene] }
+
 	const args = {
 		genome: q.genome,
 		dslabel: q.dslabel,
@@ -478,10 +484,19 @@ async function trigger_getDefaultGeneExpBins(q, ds, res) {
 		genes: [{ gene: tw.term.gene }]
 	}
 	const data = await ds.queries.geneExpression.get(args)
-	let defaultBins
-	if (ds.queries.geneExpression.gene2bins[tw.term.gene])
-		defaultBins = { default: ds.queries.geneExpression.gene2bins[tw.term.gene] }
-	else defaultBins = { default: getDefaultGeneExpBins(ds, tw, data) }
-
-	res.send({ defaultBins })
+	const lst = []
+	let min = Infinity
+	let max = -Infinity
+	for (const sampleId in data.gene2sample2value.get(tw.term.gene)) {
+		const values = data.gene2sample2value.get(tw.term.gene)
+		const value = Number(values[sampleId])
+		if (value < min) min = value
+		if (value > max) max = value
+		lst.push(value)
+	}
+	let binconfig = initBinConfig(lst)
+	binconfig.lst = bins.compute_bins(binconfig, () => {
+		return { min, max }
+	})
+	res.send({ default: binconfig })
 }
